ANDROID: mm/slab: Use EXPORT_SYMBOL for android_kmalloc_64_create
Commit 8a7073c1061a ("ANDROID: mm/slab: Create 64-byte slab caches if the
hardware supports it") introduced the android_kmalloc_64_create variable
and exported it to modules through the EXPORT_SYMBOL_GPL() macro.

Since android_kmalloc_64_create is referred to by kmalloc(), which is
inlined to all callers, including modules, it makes it so that all
modules that invoke kmalloc() have to be GPL licensed. This diverges
from upstream Linux's intent of letting kmalloc() be invoked by non-GPL
licensed modules, which is not correct. Thus, use the EXPORT_SYMBOL()
macro for exporting android_kmalloc_64_create variable, instead of
EXPORT_SYMBOL_GPL().

Fixes: 8a7073c1061a ("ANDROID: mm/slab: Create 64-byte slab caches if the hardware supports it")
Bug: 259498503
Change-Id: I96d20ec9cc17853a02e9343edb2829d0a5185f0e
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>
Signed-off-by: SirRGB <sirrgb@proton.me>